make backup and restore commands more k8s like
