--- cmd/kots/cli/restore-ls.go
@@ -1,41 +1,0 @@
-package cli
-
-import (
-	"github.com/pkg/errors"
-	"github.com/replicatedhq/kots/pkg/print"
-	"github.com/replicatedhq/kots/pkg/snapshot"
-	"github.com/spf13/cobra"
-	"github.com/spf13/viper"
-)
-
-func RestoreListCmd() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:           "ls",
-		Short:         "List available restores",
-		Long:          ``,
-		SilenceUsage:  true,
-		SilenceErrors: false,
-		PreRun: func(cmd *cobra.Command, args []string) {
-			viper.BindPFlags(cmd.Flags())
-		},
-		RunE: func(cmd *cobra.Command, args []string) error {
-			v := viper.GetViper()
-
-			options := snapshot.ListInstanceRestoresOptions{
-				Namespace: v.GetString("namespace"),
-			}
-			restores, err := snapshot.ListInstanceRestores(options)
-			if err != nil {
-				return errors.Wrap(err, "failed to list instance restores")
-			}
-
-			print.Restores(restores)
-
-			return nil
-		},
-	}
-
-	cmd.Flags().StringP("namespace", "n", "", "filter by the namespace in which kots/kotsadm is installed")
-
-	return cmd
-}
